Fix na UI, no map dos comentários e na renderização do input de add comentários
